LSTM for one emotion
As part of my work, I am developing a sentiment analysis model based on neural networks. I use the approach of creating a separate model for each emotion out of a set of 27 different emotions, with each emotion receiving a binary classification indicating whether it appears in the text or not. The use of Word2Vec and FastText provides word embeddings that allow the model to better understand the relationships and context between words. Additionally, I focus on analyzing the text sequence to ensure that the model takes into account the context of each word rather than just treating them individually.

The main challenge in this work is the severe class imbalance, as some emotions are significantly more frequent than others. To address this challenge, I use Focal Loss and Weighted Random Sampler, which help the model learn more effectively while dealing with class imbalance. I also apply Dropout regularization to prevent overfitting and improve the generalization of the model. Furthermore, I am testing the impact of various features such as emojis, advanced text frequency metrics, and sentiment lexicons on model accuracy. Ultimately, I plan to create an ensemble of networks for each emotion to allow more accurate classification of emotions in ambiguous texts.